* 1977_mcedit_type_accuracy:
  Avoid compiler errors about comparsion signed and unsigned variables
  (is_aligned_on_a_tab): return gboolean instead of int. Minor optimization.
  Use off_t instead of long for all variables related to position in file:
  Use long for line numbers and columns.
  (edit_drop_hotkey_menu): return gboolean instead of int.
  (edit_translate_keys): return gboolean instead of int.
  (edit_find_word_start): return gboolean instead of int.
  (edit_save_cmd): return gboolean instead of int.
  (edit_copy_to_X_buf_cmd): return gboolean instead of int.
  (edit_save_block): return gboolean instead of int.
  (editcmd_dialog_raw_key_query): take gboolean instead of int.
  (edit_mark_cmd): take gboolean instead of int.
  WEdit::caches_valid: change int to gboolean.
  (is_in_indent): refactoring, return gboolean instead of int.
  Ticket #1977: type accuracy of WEdit members and mceditor functions.

+++ src/editor/syntax.c
@@ -549,7 +549,7 @@ apply_rules_going_right (WEdit * edit, long i, struct syntax_rule rule)
 /* --------------------------------------------------------------------------------------------- */
 
 static struct syntax_rule
-edit_get_rule (WEdit * edit, long byte_index)
+edit_get_rule (WEdit * edit, off_t byte_index)
 {
     long i;
 
@@ -1258,7 +1258,7 @@ edit_read_syntax_file (WEdit * edit, char ***pnames, const char *syntax_file,
 #define NENTRIES 30
     FILE *f, *g = NULL;
     char *args[1024], *l = NULL;
-    int line = 0;
+    long line = 0;
     int result = 0;
     int count = 0;
     char *lib_file;
@@ -1415,7 +1415,7 @@ get_first_editor_line (WEdit * edit)
 /* --------------------------------------------------------------------------------------------- */
 
 void
-edit_get_syntax_color (WEdit * edit, long byte_index, int *color)
+edit_get_syntax_color (WEdit * edit, off_t byte_index, int *color)
 {
     if (!tty_use_colors ())
         *color = 0;
